refactor: useTimer 훅 분리
