+++ src/pages/Tools/Timer/index.tsx
@@ -1,35 +1,21 @@
-import { useEffect, useState } from 'react';
 import { FaPause, FaPlay } from 'react-icons/fa6';
+
+import useTimer from '@Hooks/tools/timer/useTimer';
 
 import * as S from './style';
 
 function Timer() {
-  const [initTime, setInitTime] = useState(320);
-  const [time, setTime] = useState(320);
-  const [state, setState] = useState<'stop' | 'play'>('stop');
+  const { state, time, progress, toggleState } = useTimer();
 
   const minute = Math.floor(time / 60);
   const second = time - minute * 60;
+
   const displayMinute = minute < 10 ? '0' + String(minute) : minute;
   const displaySecond = second < 10 ? '0' + String(second) : second;
+
   const displayTime = `${displayMinute}:${displaySecond}`;
+
   const timerButton = state === 'stop' ? <FaPlay /> : <FaPause />;
-  const progress = ((initTime - time) / initTime) * 100;
-
-  const handleClickTimerButton = () => {
-    if (state === 'play') setState('stop');
-    else setState('play');
-  };
-
-  useEffect(() => {
-    if (time === 0 || state === 'stop') return;
-
-    const timer = setInterval(() => {
-      setTime((prev) => prev - 1);
-    }, 1000);
-
-    return () => clearInterval(timer);
-  }, [time, state]);
 
   return (
     <S.Layout>
@@ -39,9 +25,7 @@ function Timer() {
         <S.TimeBar>
           <S.ProgressBar progress={progress}></S.ProgressBar>
         </S.TimeBar>
-        <S.TimerButton onClick={handleClickTimerButton}>
-          {timerButton}
-        </S.TimerButton>
+        <S.TimerButton onClick={toggleState}>{timerButton}</S.TimerButton>
       </S.TimeContainer>
     </S.Layout>
   );
